fix: wording and abbreviations
